fix: move mt-floating-ui out of _internal folder
