+++ packages/component-library/src/components/mt-floating-ui/mt-floating-ui.mdx
@@ -1,7 +1,7 @@
 import { ArgTypes, Canvas, Meta } from "@storybook/blocks";
 
 import * as FloatingUiStories from "./mt-floating-ui.stories";
-import StorybookPageHeader from "../../../docs/_components/storybook-page-header.js";
+import StorybookPageHeader from "../../docs/_components/storybook-page-header.js";
 
 <Meta of={FloatingUiStories} />
 
@@ -9,7 +9,7 @@ import StorybookPageHeader from "../../../docs/_components/storybook-page-header
   title="Floating UI"
   tagName="mt-floating-ui"
   packageImports="MtFloatingUi"
-  sourcePath="packages/component-library/src/components/_internal/mt-floating-ui"
+  sourcePath="packages/component-library/src/components/mt-floating-ui"
 >
   **Floating UI** is the low-level primitive for building custom popover-style
   elements with automatic positioning. Use it when you need floating content
