+++ packages/component-library/src/components/mt-popover/mt-popover.vue
@@ -52,7 +52,7 @@ import MtIcon from "../mt-icon/mt-icon.vue";
 import MtPopoverItem from "../mt-popover-item/mt-popover-item.vue";
 import MtButton from "../mt-button/mt-button.vue";
 import MtSmoothReflow from "../_internal/mt-smooth-reflow.vue";
-import MtFloatingUi from "../_internal/mt-floating-ui/mt-floating-ui.vue";
+import MtFloatingUi from "../mt-floating-ui/mt-floating-ui.vue";
 import type { TranslateResult } from "vue-i18n";
 import type { View } from "./mt-popover.interfaces";
 import MtText from "@/components/mt-text/mt-text.vue";
